requirements.txt
requests
python-dotenv
python-telegram-bot>=20.0                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    